// src/app/models/resumen-x-grupo.model.ts
/**
 * Utilizado en dashboard abasto > Existencias (CPMs) > Existencias X Grupo
 */
export class ResumenXGrupo {
    key: string;
    municipio: string;
    clues: string;
    nombreUnidad: string;
    nivelAtencion: string;
    tipologia: string;
    categoria: string;
    clavesManejadas: number;
    clavesDesabasto: number;
    porcentajeDesabasto: number;
}